♻️ New sidebar styles for streamline
