+++ static/app/views/issueDetails/groupEventDetails/groupEventDetails.tsx
@@ -28,7 +28,7 @@ import {useSyncedLocalStorageState} from 'sentry/utils/useSyncedLocalStorageStat
 import GroupEventDetailsContent from 'sentry/views/issueDetails/groupEventDetails/groupEventDetailsContent';
 import GroupEventHeader from 'sentry/views/issueDetails/groupEventHeader';
 import GroupSidebar from 'sentry/views/issueDetails/groupSidebar';
-import StreamlinedSidebar from 'sentry/views/issueDetails/streamline/sidebar';
+import StreamlinedSidebar from 'sentry/views/issueDetails/streamline/sidebar/sidebar';
 
 import ReprocessingProgress from '../reprocessingProgress';
 import {
