+++ static/app/views/issueDetails/streamline/sidebar/streamlinedExternalIssueList.spec.tsx
